Movied Kernel to Higher-Half (#16)
* supressed warnings. Attempted higher-half relocation

* set control registers after paging enabled

* removed first page entry

+++ kernel/mm/paging.c
@@ -28,7 +28,7 @@ void init_paging()
     }
 
     // assign table as only entry in directory
-    page_directory[0] = page_table;
+    page_directory[0] = (uint32_t) page_table;
 
 }
